requirements.txt
django
factory_boy